app/requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
ufal.udpipe==1.2.0.3
Werkzeug==2.0.2
wget==3.2
gunicorn>=19,<20F
